Refactor path_generator
All parsing code has been moved into the parse_docket_id method.

All KeyErrors have been replaced by instead checking if
the key exists in the first place.

Created a function that checks if a series of nested keys exists.

get_attributes() now returns a value of 'unknown' if the attribute
could not be found or parsed in the JSON.

get_attributes() now logs whenever it parses docket_id from item_id.

Changed the name of get_attributes to _get_attributes since it is a
private method to stay consistent with python conventions.